boundaries/marketing/referral/requirements.txt
dependency-injector @ git+https://github.com/shortlink-org/python-dependency-injector@master
uvicorn
redis
protobuf
loguru
quart
sqlalchemy
python-dotenv
setuptools>=69.1.1 # not directly required, pinned by Snyk to avoid a vulnerability

# Opentelemetry
pyroscope-io==0.8.5
opentelemetry-api==1.22.0
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-sqlalchemy

# For testing:
pytest
allure-pytest
pytest-asyncio
pytest-cov
pytest-factoryboy
httpx
testcontainers

# For code quality:
ruff

# For profiling:
scalene
werkzeug>=3.0.1 # not directly required, pinned by Snyk to avoid a vulnerability
